CLEAN CODE!
