+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod brew;
 pub mod cli;
 pub mod commands;
 pub mod config;
